Rewrite certificate handling
Private keys should (if possible) only exist on the system they are needed on.
Therefore the creation of certificates is now split:
* The private key is created on the host they are used on
* The CSR is created on the host and copied to the ansible controller
* The user is guided to create a certificate from the CSR

The certificate strings are removed from the inventory and example vaults

All certificates and similiar files are stored in paths which are defined as vars to make the usage more consistent

The root and intermediate certifcates are stored in files to make the handling easier

Two playbooks are created for easier testing:
create_ownca.yml creates a selfsigned ca which is used to sign the CSR inside the selfsign_csr.yml playbook.